feat: Add shuffle button to meme gallery
Adds a refresh button to the meme gallery screen. Clicking this button shuffles the order of the meme templates in the grid.

The list of meme templates has been refactored into a `MemeTemplates` object with a mutable state to allow for shuffling.

Additionally, the drawable resource filenames for meme templates have been renamed to remove leading zeros.